pull music type search into individual component for better breakdown of code
